+++ src/main/java/source/usecases/dto/request/clothes/ClothesUpdateRequestModel.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Set;
 
@@ -30,5 +31,5 @@ public class ClothesUpdateRequestModel {
 
     private String comment;
 
-    private Float satisfaction;
+    private BigDecimal satisfaction;
 }
